% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{config_set}
\alias{config_get}
\title{Write configurations to _deploy.yml or DESCRIPTION}
\usage{
config_set(..., .in = c("deploy", "desc"))

config_get(.in = c("default", "desc", "deploy"))
}
\arguments{
\item{...}{Named values to write to the configuration}

\item{.in}{The location to write the configuration to or read it from. For
\code{config_get()} multiple values are allowed and the values from the sources
will be merged together according to the priority in the description.}
}
\value{
This function is called for its side effect
}
\description{
deployr execution can be controlled through the use of a \verb{_deploy.yml} file
in the projects root or by setting the same values in the \code{Config/deploy}
field of the projects \code{DESCRIPTION} file. The priority of settings are
\verb{_deploy.yml} > \code{DESCRIPTION} > defaults. \code{config_set()} allow you to write
configuration values to either of these places. \code{config_get} retrieves the
configuration for the current project (by default merging all sources
together).
}
\details{
\subsection{Recognised settings}{

The following list provides an overview of the settings that are currently
used by deployr:
\itemize{
\item \code{distDir}: The location of files generated by static projects. Defaults to
\verb{./dist/}.
\item \code{port}: The port to use for dynamic projects. Defaults to \code{50505}.
\item \code{host}: The host to use for dynamic projects. Defaults to \code{"127.0.0.1"}.
\item \code{workerId}: An ID that identifies the specific worker running the code.
Will be used by Shiny runtimes to correctly route clients to the same
worker. Defaults to \code{""}.
\item \code{sharedSecret}: A shared secret used by Plumber, plumber2 and Shiny as a
simple authentication mechanism. Defaults to \code{NULL}.
\item \code{logger}: The name of the R package used for logging by the server.
Currently unused. Defaults to \code{"logger"}.
\item \code{staticFormat}: The format to render to for static projects. Is passed to
\code{output_format} in quarto and rmarkdown rendering pipelines. Defaults to
\code{NULL}
\item \code{params}: A list of parameters to overwrite in the YAML front-matter of
quarto and rmarkdown documents. Will be passed to
\code{quarto::quarto_render(execute_params)} and \code{rmarkdown::render(params)}.
Defaults to \code{list()}
\item \code{quartoArgs}: Additional arguments for quarto, passed to
\code{quarto::quarto_render(quarto_args)}. Defaults to \code{NULL}
\item \code{pandocArgs}: Additional arguments for quarto, passed to
\code{quarto::quarto_render(pandoc_args)}. Defaults to \code{NULL}
\item \code{hugoArgs}: Additional arguments for blogdown, passed to
\code{blogdown::build_site(hugo_args)}. Defaults to \code{NULL}
}
}
}
