% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create skeleton for a new project}
\usage{
create(
  type,
  title = "default",
  message_fun = function(...) NULL,
  file_creator = function(...) NULL
)
}
\arguments{
\item{type}{The project type to create. See \code{\link[=project_category]{project_category()}} for an
overview of the different possible types}

\item{title}{The title of the project. Will be used in various project types
to fill in titles}

\item{message_fun}{A function (usually \code{cli::cli_bullets()}) used to create
status messages. Passed in as an argument to avoid additional dependencies
for dployr}

\item{file_creator}{A function (usually \verb{\\(...) fs::file_create(fs::path(...))})
used to create files. Passed in as an argument to acoid additional
dependencies}
}
\value{
A character vector giving the package dependencies for the project
}
\description{
This function sets up the \verb{exec/} folder according to the project type and
returns the names of the packages the project type requires. It is not
intended to be called directly but instead orchestrated by
\code{producethis::create_production()}
}
\keyword{internal}
