% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dployr-package.R
\docType{package}
\name{dployr-package}
\alias{dployr}
\alias{dployr-package}
\title{dployr: Runtime Dependencies for 'producethis'}
\description{
Provides a low-dependency package for the parts of 'producethis' that are required at runtime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/dployr}
  \item Report bugs at \url{https://github.com/thomasp85/dployr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
