% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{execute}
\alias{execute}
\title{Execute a producethis project}
\usage{
execute()
}
\value{
This function is called for its side effects
}
\description{
This function assumes the current working directory is the root of a project
created according to producethis conventions. If this is the case it will
execute the project according to its type and the content in \code{./exec}
}
